moved default_value.go to internal/value dir
